Supports "package" variable in Cypress runs
