+++ cypress.ci.json
@@ -1,0 +1,3 @@
+{
+  "fileServerFolder": "{PACKAGE}/storybook-static"
+}
