+++ cypress.json
@@ -1,5 +1,5 @@
 {
-  "integrationFolder": "../../examples",
+  "integrationFolder": "./examples",
   "testFiles": "**/*.test.js",
   "viewportWidth": 500
 }
