+++ cypress/support/commands/assertAreas.js
