--- packages/atomic-layout/cypress.ci.json
@@ -1,3 +1,0 @@
-{
-  "fileServerFolder": "storybook-static"
-}
