+++ packages/atomic-layout/cypress/plugins/index.js
@@ -1,50 +1,17 @@
-const fs = require('fs')
-const webpackPreprocessor = require('@cypress/webpack-preprocessor')
-const babelConfig = require('../../babel.config')
+// ***********************************************************
+// This example plugins/index.js can be used to load plugins
+//
+// You can change the location of this file or turn off loading
+// the plugins file with the 'pluginsFile' configuration option.
+//
+// You can read more here:
+// https://on.cypress.io/plugins-guide
+// ***********************************************************
 
-const webpackOptions = {
-  module: {
-    rules: [
-      {
-        test: /\.(j|t)sx?$/,
-        exclude: /node_modules/,
-        use: [
-          {
-            loader: 'babel-loader',
-            options: babelConfig,
-          },
-        ],
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.ts', '.tsx', '.spec.jsx', '.spec.js', '.jsx', '.js'],
-  },
-}
-
-const getCypressConfig = (envName = '') => {
-  const configFilename = ['cypress', envName, 'json'].filter(Boolean).join('.')
-
-  const baseConfig = require('../../cypress.json')
-  const envConfig = envName ? require(`../../cypress.${envName}.json`) : {}
-
-  console.log(`Loading base Cypress config...`)
-
-  if (envName) {
-    console.log(`Loading environmental Cypress config: ${configFilename}...`)
-  }
-
-  return Object.assign({}, baseConfig, envConfig)
-}
+// This function is called when a project is opened or re-opened (e.g. due to
+// the project's config changing)
 
 module.exports = (on, config) => {
-  on(
-    'file:preprocessor',
-    webpackPreprocessor({
-      webpackOptions,
-      watchOptions: {},
-    }),
-  )
-
-  return getCypressConfig(config.env.envName)
+  // `on` is used to hook into various events Cypress emits
+  // `config` is the resolved Cypress config
 }
